Begin cleaning up member initialization lists.
Using member initialization lists instead of assignment
within the constructor is recommended c++ practice.

This is also supposed to allow the compiler to generate
better code, which seems to be the case for g++.
The constructors in this patch shrink by about 64 and
128 bytes, respectively.  If more than one version of a
particular constructor is generated, then this savings seems
to apply to each copy.

Adopt a formatting convention of one initializer per line,
indented by 4 spaces, sorted by member declaration order
within the class to avoid g++ warnings.

Per item 4 of 'Effective C++', ensure that all member
initializers are explicitly invoked, even when the compiler
would do so automatically.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>


git-svn-id: https://www.taskjuggler.org/svn/taskjuggler/trunk@1402 e1914e07-63f8-0310-9059-d6d858d7cdca