Patch from Andreas: Cleanup includes
Attached is a patch file that reduces TaskJuggler's include hierarchy according to the two rules

(1) Do not include headerfiles redundantly in your module interface (headerfile), when they are coming with a headerfile that your module interface requires.

(2) A module (cpp file) shall include its corresponding headerfile as the first significant line of code in order to assure that all interface definitions compile stand-alone.

This results in the removal of a significant number of #include directives from various headerfiles, and in positional shifts of #include directives in several implementation files.

I have not weeded through the implementation files and removed redundant #include directives therein just yet.


git-svn-id: https://www.taskjuggler.org/svn/taskjuggler/trunk@1408 e1914e07-63f8-0310-9059-d6d858d7cdca