tj-graphviz-pert: rewritten dot stylesheet to have same look as with fdp
The dot stylesheet used to draw supertasks as regular nodes linked to their
subtasks, which was not really usefull due to a graphviz limitation (dot does
not support edges between nodes and clusters or between clusters).

Now the dot stylesheet produces the same kind of appearance than the fdp one
(in fact, it looks far better), but it uses awful hacks using "ltail" and
"lhead" graphviz attributes with one invisible node by cluster which is the
source or target of an edge where the cluster should be this source or target.