Removing redundant error checking related to booking declaration.
Both checks are redundant with those made in Task::preScheduleOk().
Moreover the check concerning milestones is less precise than the
one in Task::preScheduleOk() because the later detects implicit
milestones (tasks with no effort, no duration and no length)
whereas the removed check only handles tasks declared as milestones
by the "milestone" property (because it is done earlier in the
pre-scheduling process).