Auto enlarge "scheduled" task time frame if a "booking" does not fit it.
For a given scenario, when a resource booking is found for a task that
has no longer to be scheduled ("scheduled" property is set for this
scenario) and does not fit the task time frame, then the time frame
is modified for the booking to fit it.

This patch makes time tracking easier since it avoids the painfull
manual work about tasks boundaries needed to avoid plenty of
"booking of resource XXX is outside of task interval" errors.

It also avoids at all the need of any property apart from "scheduled"
in such a task, including: "start", "end", "depends", "precedes",
"effort", "duration", "length", "allocate". An historical task can
now be almost empty, being only a support for bookings.