Adding ratio-based usage limits for "weeklymax" and "monthlymax".
When used in "resource" or "project" context (but not in "allocate"
context), "weeklymax" and "monthlymax" limits now accept a ratio
parameter, that is, a number > 0 and <= 1 without unit, which
define a limit as a ratio of the real working time in the period,
not including vacation.

This is usefull for resources that are shared among several
projects, to accurately handle general and personal vacations of
the resource.

For instance a resource working 50% of his/her time on the project
was previously handled with limits such as "weeklymax 2.5d".
This had huge drawbacks for weeks with less than 5 days of
workable time. For instance a 4-days week with, let say, a day
off due to national hollyday will make the resource work 63% of
his/her time instead of 50% (2.5d/4d).
Moreover, when the same person come back from beach on thursday,
he/she will even work 100% of his/her time, since 2 days are
less than 2.5.
Defining the limit with "weeklymax 0.5" does not have this
drawback, this will be interpreted as "2d" on 4-days weeks, and
as "1d" on 2-days weeks.