Separating "over resource limits load" from free load in graphical charts.
Until now, the time that is not allocatable to a resource due to
its usage limits was reported as free load with green bars on Gantt
and resource charts. This is misleading or at less hard to read for
part time resources.

Now this time is reported separatly with another bar in another
color.