Fixed a rare scheduler bug involving containers early boundary date propagation
In a rare case described in the example below, the scheduler could not
find a way to plan a task with several dependencies, ending with the
following error message:
Impossible dependency
Task 'T1' ends at 2009-1-10 23:59:59 but needs to precede
task 'M1' which has a 'plan' start time of 2009-1-10 00:00:00

This was due to the fact that, in ASAP scheduling, when the end date
of a container (C0) is computed from a follower (M1) start date, it
was propagated to all of its other followers with no known start date
(T1, M2). However at this point the container computed end date can
still change, since its subtasks are not yet planed.
If one (T1) of the container followers is too a predecessor of the
task from which the temporary end date was taken (M1), then it (T1)
cannot be scheduled since its start date is equal to one of its
followers start date.

Task::propagateEnd() now skips end date to start date propagation to
followers for containers which don't yet have their final end date.

The same thing apply to ALAP scheduling, in Task::propagateStart()
for start date to end date propagation to predecessors.

Here is the sample project file used to reproduce the bug:

project prj "My Project" "1.0" 2009-1-1 - 2009-1-31

task C0 "C0" {
  task T0 "T0" {
    duration 6d
  }
  # the bug does not occur if C0 is a leaf, but only if
  # it is a container
  # fix: avoid propagating container dates when they are
  # not stabilized (when children are not all scheduled)
  # however this may make some projects unschedulable whereas
  # they are schedulable now
}

task T1 "T1" {
  # bug: this task cannot be scheduled whereas a dumb human
  # would be able to do so
  # it seems that M1 date is propagated to C0 through the
  # "depends" constraint, and then that it is propagated
  # again to all C0 followers, including T1, which is a
  # problem since it then uses the start date for ASAP
  # scheduling (hard constraint) and therefore is not able
  # to comply with it end soft constraint ("depends" from
  # M1 too).
  depends C0
  duration 1d
  # uncommenting following line is a workaround to the bug
  # and it allows looking at M2 scheduling as explained
  # below
  #start 2009-1-7
}

task M1 "M1" {
  milestone
  start 2009-1-10
  depends T1
  # the bug does not occur if "depends T1" is replaced with
  # a "precedes M1" in T1, probably because of the switch to
  # ALAP scheduling that this implies
  depends C0
  # the bug does not occur without C0 dependency, it
  # probably has something to do with the fact that C0
  # dependency is induced due to C0->T1->M1 dependencies
  # note: another possible fix could have been to avoid
  # following induced dependencies
}

task M2 "M2" {
  # the bug still occur without M2, but M2 is delayed
  # the same way than T1, which is interesting too
  # for understanding the problem
  # M2 start date is always the same than T1's whatever
  # date is set using "date" in the T1 section
  milestone
  depends C0
}

taskreport "Gantt" {
  columns name,start,end,chart
  sorttasks tree,sequenceup
}